completed implement

+++ emart-microservice/seller-center-service/src/main/java/com/fsd/emart/seller/bean/SalesOverviewInfo.java
@@ -2,15 +2,16 @@ package com.fsd.emart.seller.bean;
 
 import java.io.Serializable;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import lombok.Data;
 
 @Data
 public class SalesOverviewInfo implements Serializable {
-	/** UID */
-	private static final long serialVersionUID = 4343367517598045502L;
+    /** UID */
+    private static final long serialVersionUID = 4343367517598045502L;
 
-	private Integer count;
-	private BigDecimal amount;
+    private BigInteger count;
+    private BigDecimal amount;
 
 }
